requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai-whisper==20231117
rapidfuzz==3.5.2
numpy==1.24.3
torch==2.0.1
torchaudio==2.0.2
ffmpeg-python==0.2.0
werkzeug==2.3.7
python-multipart==0.0.6
librosa==0.10.1
soundfile==0.12.1